Split storage.js into dataset-clients and sync-updates
